Add timed/arcade modes, cascade clearing and high scores to the C++ side
The last three Python-only pieces of the game, ported and graded:

- Sim grows the three game modes: timed's five-minute clock with its
  closing score multiplier and two-step expiry, arcade's level ramp,
  gravity curve and rising garbage at every tier - with the garbage
  holes dealt in from outside, so the sim still never rolls dice.
- All three clear styles. The sticky and linked cascades shipped broken
  upstream (out-of-bounds fills, stale fill seeds that hung the game on
  a cleared garbage row, sticky groups stuck by fiat, a blocked shape
  re-queued forever); engine/shapes.py is repaired first, pinned by the
  new tools/test_cascade.py, and the C++ port is graded against the
  repaired behaviour - including the movement loop's settled-collision
  verdict, which no natural game reaches, pinned by hand-built boards
  on both sides (test_cascade.py and the new cascade_check ctest).
- The high score table: the same 720-byte data/hiscore.dat, byte for
  byte, backup mirror, repair ladder and tie quirks included, graded by
  the new hiscore_cross test - 17 scripted submissions through both
  codecs must produce identical files and announce identical places.
- The GUI grows Play timed / Play arcade, the High scores screen with
  the Python game's three pages, the game-over name entry, and a Line
  clears setting; the timer stat counts down in timed games and a level
  stat joins the roster.
- Trace coverage rises to thirty scripted games: every arcade garbage
  tier, both cascade styles over rubble and under garbage, and two timed
  games whose clock dies with a clear still resolving - which pin the
  loss ordering (gameover before the frame's clear cue, high score and
  replay taken before the final clear lands its points).
- The sim's score goes 64-bit end to end (sim, replay, traces): Python's
  integer is unbounded, and a marathon must not wrap where the reference
  keeps counting. The live VS stat now counts dug garbage as downstack,
  and the clear banner names the last chain link with the '{} LINES'
  fallback, both matching the Python game.
- engine/game.py: fix a crash in arcade's garbage spawner (levels 64-127
  read a misspelled attribute); engine/filehandler.py honours
  FORCETRIS_HISCORE so tests keep out of the real table.

+++ cpp/include/forcetris/replay.hpp
@@ -42,7 +42,8 @@ struct Meta {
 	int spinrule = 0;
 	int cleartype = 0;
 	int das = 0, arr = 0, dcd = 0, sdf = 0, are = 0;
-	int score = 0;
+	// 64-bit, like the sim's own score: Python's integer is unbounded.
+	long long score = 0;
 	int lines = 0;
 	int downstack = 0;
 	double seconds = 0.;
@@ -64,7 +65,7 @@ struct Placement {
 	bool perfect = false;
 	int combo = 0;           // As shown on the HUD: counter minus one.
 	int b2b = 0;
-	int score = 0;
+	long long score = 0;
 	double elapsed = 0.;
 	std::vector<std::string> rows;
 	std::vector<int> queue;  // The previews the player could see. Empty in v2.
@@ -94,7 +95,7 @@ struct Summary {
 	double ppp = 0.;
 	double pps = 0.;
 	int lines = 0;
-	int score = 0;
+	long long score = 0;
 	double seconds = 0.;
 	std::map<int, int> clears;
 	int spins = 0;
@@ -138,7 +139,7 @@ public:
 	// Stamp the totals and hand back the replay, or nothing if the game was
 	// too short to keep.
 	std::optional<Replay> finish (
-		int score, int lines, int downstack, double seconds);
+		long long score, int lines, int downstack, double seconds);
 
 private:
 	Meta meta_;
